[Refactor] 디렉토리구조 개편으로 인한 테스트코드 import 수정
