+++ src/test/java/kr/KWGraduate/BookPharmacy/jwt/JWTUtilTest.java
@@ -1,6 +1,7 @@
 package kr.KWGraduate.BookPharmacy.jwt;
 
-import kr.KWGraduate.BookPharmacy.dto.token.TokenDto;
+import kr.KWGraduate.BookPharmacy.global.security.common.dto.TokenDto;
+import kr.KWGraduate.BookPharmacy.global.security.common.util.JWTUtil;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
