+++ src/test/java/kr/KWGraduate/BookPharmacy/service/KeywordServiceTest.java
@@ -1,22 +1,17 @@
 package kr.KWGraduate.BookPharmacy.service;
 
-import kr.KWGraduate.BookPharmacy.entity.*;
-import kr.KWGraduate.BookPharmacy.repository.BookRepository;
-import kr.KWGraduate.BookPharmacy.repository.ClientRepository;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
+import kr.KWGraduate.BookPharmacy.domain.book.repository.BookRepository;
+import kr.KWGraduate.BookPharmacy.domain.client.repository.ClientRepository;
+import kr.KWGraduate.BookPharmacy.domain.keyworditem.service.KeywordItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @SpringBootTest
 @Transactional
 class KeywordServiceTest {
     @Autowired
-    private KeywordService keywordService;
+    private KeywordItemService keywordService;
 
     @Autowired
     private BookRepository bookRepository;
