+++ src/test/java/kr/KWGraduate/BookPharmacy/service/OneLinePrescriptionServiceTest.java
@@ -2,13 +2,16 @@ package kr.KWGraduate.BookPharmacy.service;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
-import kr.KWGraduate.BookPharmacy.dto.client.ClientDetails;
-import kr.KWGraduate.BookPharmacy.dto.oneLinePrescription.request.OneLineCreateDto;
-import kr.KWGraduate.BookPharmacy.dto.oneLinePrescription.request.OneLineUpdateDto;
-import kr.KWGraduate.BookPharmacy.entity.*;
-import kr.KWGraduate.BookPharmacy.repository.BookRepository;
-import kr.KWGraduate.BookPharmacy.repository.ClientRepository;
-import kr.KWGraduate.BookPharmacy.repository.OneLinePrescriptionRepository;
+import kr.KWGraduate.BookPharmacy.domain.book.domain.Book;
+import kr.KWGraduate.BookPharmacy.domain.book.repository.BookRepository;
+import kr.KWGraduate.BookPharmacy.domain.client.domain.Client;
+import kr.KWGraduate.BookPharmacy.domain.client.repository.ClientRepository;
+import kr.KWGraduate.BookPharmacy.domain.onelineprescription.domain.OneLinePrescription;
+import kr.KWGraduate.BookPharmacy.domain.onelineprescription.dto.request.OneLineCreateDto;
+import kr.KWGraduate.BookPharmacy.domain.onelineprescription.dto.request.OneLineUpdateDto;
+import kr.KWGraduate.BookPharmacy.domain.onelineprescription.repository.OneLinePrescriptionRepository;
+import kr.KWGraduate.BookPharmacy.domain.onelineprescription.service.OneLinePrescriptionService;
+import kr.KWGraduate.BookPharmacy.global.security.auth.dto.ClientDetails;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
